move getPod and getMigration
